+++ mtools/mlogfilter/filters/base_filter.py
@@ -25,11 +25,6 @@ class BaseFilter(object):
         """
         pass
 
-    def teardown(self):
-        """ hook to teardown anything necessary for the filter at completion
-        """
-        pass
-
     def accept(self, logevent):
         """ overwrite this method in subclass and return True if the provided 
             logevent should be accepted (causing output), or False if not.