+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes to mtools
 #### version 1.1.6dev
 
 #### version 1.1.5
+  * Changed logline filter to match the accept connection line for a thread(#218)
 
   * added workaround for compile errors with XCode 5.1 / clang 3.4 (#203)
   * mlaunch: fixed bug when using `--binarypath` and passing arguments through to mongod/mongos (#217)